Feature/NN-810 paymaart  customer android filter all transactions

Approved-by: Sreekuttan J

+++ app/src/main/java/com/afrimax/paymaart/data/model/TransactionDetailsResponse.kt
@@ -13,7 +13,7 @@ data class TransactionDetailsResponse(
 
 data class TransactionDetails(
     @SerializedName("grossTransactionFee")
-    val grossTransactionFee: Int,
+    val grossTransactionFee: Double,
     @SerializedName("netTransactionFee")
     val netTransactionFee: Double,
     @SerializedName("vatAmount")
